tests and ros2

+++ src/controllers/ros2/deleteAction.ts
@@ -7,6 +7,8 @@ import { BadRequest, Forbidden } from '../../errors/bad-request';
 import Organization, { OrganizationPermissions } from '../../models/Organization';
 import ROS2ActionModel from '../../models/Ros2/Ros2Action';
 import RobotPart from '../../models/RobotPart';
+import Robot from '../../models/Robot';
+import Ros2SystemModel from '../../models/Ros2/Ros2System';
 
 interface DeleteActionParams {
     robotId: string
@@ -30,8 +32,14 @@ const deleteAction: RequestHandler<any> = async (req: Request<DeleteActionParams
       throw new Forbidden('User do not have the authorization for deleting ros2 related protocol');
     };
 
-    const part = await RobotPart.findOne({ _id: params.partId });
+    const robot = await Robot.findById(params.robotId);
+    if (!robot) { throw new BadRequest('The robot does not exist'); };
+    const part = robot.parts.find(e => e._id.toString() === params.partId);
     if (!part) { throw new BadRequest('The part does not exist'); };
+
+    const ros2System = await Ros2SystemModel.getByRobotId(robot.id);
+    ros2System.actions = ros2System.actions.filter(e => e._id.toString() !== params.actionId);
+    await ros2System.save();
 
     await ROS2ActionModel.deleteOne({ _id: params.actionId });
