+++ src/controllers/ros2/deletePublisher.ts
@@ -7,6 +7,8 @@ import { BadRequest, Forbidden } from '../../errors/bad-request';
 import Organization, { OrganizationPermissions } from '../../models/Organization';
 import ROS2PublisherModel from '../../models/Ros2/Ros2Publisher';
 import RobotPart from '../../models/RobotPart';
+import Robot from '../../models/Robot';
+import Ros2SystemModel from '../../models/Ros2/Ros2System';
 
 interface DeletePublisherParams {
     robotId: string
@@ -16,7 +18,8 @@ interface DeletePublisherParams {
 
 const deletePublisherSchemaParams = Joi.object<DeletePublisherParams>().keys({
   robotId: Joi.string().required(),
-  partId: Joi.string().required()
+  partId: Joi.string().required(),
+  publisherId: Joi.string().required()
 });
 
 const deletePublisher: RequestHandler<any> = async (req: Request<DeletePublisherParams, {}, {}>, res, next) => {
@@ -29,8 +32,14 @@ const deletePublisher: RequestHandler<any> = async (req: Request<DeletePublisher
       throw new Forbidden('User do not have the authorization for deleting ros2 related protocol');
     };
 
-    const part = await RobotPart.findOne({ _id: params.partId });
+    const robot = await Robot.findById(params.robotId);
+    if (!robot) { throw new BadRequest('The robot does not exist'); };
+    const part = robot.parts.find(e => e._id.toString() === params.partId);
     if (!part) { throw new BadRequest('The part does not exist'); };
+
+    const ros2System = await Ros2SystemModel.getByRobotId(robot.id);
+    ros2System.publishers = ros2System.publishers.filter(e => e._id.toString() !== params.publisherId);
+    await ros2System.save();
 
     await ROS2PublisherModel.deleteOne({ _id: params.publisherId });
