+++ src/controllers/ros2/deleteTopic.ts
@@ -7,6 +7,8 @@ import { BadRequest, Forbidden } from '../../errors/bad-request';
 import Organization, { OrganizationPermissions } from '../../models/Organization';
 import RobotPart from '../../models/RobotPart';
 import ROS2TopicModel from '../../models/Ros2/Ros2Topic';
+import Robot from '../../models/Robot';
+import Ros2SystemModel from '../../models/Ros2/Ros2System';
 
 interface DeleteTopicParams {
     robotId: string
@@ -16,7 +18,8 @@ interface DeleteTopicParams {
 
 const deleteTopicSchemaParams = Joi.object<DeleteTopicParams>().keys({
   robotId: Joi.string().required(),
-  partId: Joi.string().required()
+  partId: Joi.string().required(),
+  topicId: Joi.string().required()
 });
 
 const deleteTopic: RequestHandler<any> = async (req: Request<DeleteTopicParams, {}, {}>, res, next) => {
@@ -29,8 +32,14 @@ const deleteTopic: RequestHandler<any> = async (req: Request<DeleteTopicParams, 
       throw new Forbidden('User do not have the authorization for deleting ros2 related protocol');
     };
 
-    const part = await RobotPart.findOne({ _id: params.partId });
+    const robot = await Robot.findById(params.robotId);
+    if (!robot) { throw new BadRequest('The robot does not exist'); };
+    const part = robot.parts.find(e => e._id.toString() === params.partId);
     if (!part) { throw new BadRequest('The part does not exist'); };
+
+    const ros2System = await Ros2SystemModel.getByRobotId(robot.id);
+    ros2System.topics = ros2System.topics.filter(e => e._id.toString() !== params.topicId);
+    await ros2System.save();
 
     await ROS2TopicModel.deleteOne({ _id: params.topicId });
